Fixed HTTP CalendarAPI responses

+++ src/main/java/com/spring/phisioweb/api/event/CalendarAPIRest.java
@@ -11,6 +11,8 @@ import com.spring.phisioweb.model.Event;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.format.annotation.DateTimeFormat.ISO;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -26,7 +28,13 @@ public class CalendarAPIRest {
     @GetMapping("/api/events")
     @JsonSerialize(using = LocalDateTimeSerializer.class)
     Iterable<Event> events(@RequestParam("start") @DateTimeFormat(iso = ISO.DATE_TIME) LocalDateTime start, @RequestParam("end") @DateTimeFormat(iso = ISO.DATE_TIME) LocalDateTime end) {
-        return eventRepo.findBetween(start, end);
+        ResponseEntity<Iterable<Event>> response = loadEvents(start, end);
+        if(response.getStatusCode() == HttpStatus.OK) {
+            Iterable<Event> listEvent = response.getBody();
+            return listEvent;
+        }else {
+            return null;
+        }
     }
 
     @PostMapping("/api/events/create")
@@ -38,18 +46,65 @@ public class CalendarAPIRest {
         event.setEnd(params.start.plusMinutes(params.servDuration));
         event.setText(params.text);
         event.setResource(params.resource);
-        eventRepo.save(event);
-        return event;
+        event.setServDuration(params.servDuration);
+        ResponseEntity<Boolean> response = checkEvent(event);
+        if(response.getStatusCode() == HttpStatus.OK) {
+            eventRepo.save(event);
+            return event;
+        }else {
+            return null;
+        }
     }
 
     @PostMapping("/api/events/delete")
     @JsonSerialize(using = LocalDateTimeSerializer.class)
     @Transactional
     EventDeleteResponse deleteEvent(@RequestBody EventDeleteParams params) {
-        eventRepo.deleteById(params.id);
-        return new EventDeleteResponse() {{
-            message = "Event deleted";
-        }};
+        ResponseEntity<Boolean> response = eventExists(params.id);
+        if(response.getStatusCode() == HttpStatus.OK) {
+            eventRepo.deleteById(params.id);
+            return new EventDeleteResponse() {{
+                message = "Event deleted";
+            }};
+        }else {
+            return null;
+        }
+    }
+
+    public ResponseEntity<Iterable<Event>> loadEvents(LocalDateTime start, LocalDateTime end) {
+        Iterable<Event> listEvents = eventRepo.findBetween(start, end);
+        if(listEvents == null) {
+            return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+        }else {
+            return ResponseEntity.status(HttpStatus.OK).body(listEvents);
+        }
+    }
+
+    public ResponseEntity<Boolean> checkEvent(Event event) {
+        if(eventIsOK(event)) {
+            return ResponseEntity.status(HttpStatus.OK).body(true);
+        }else {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+        }
+    }
+
+    //Method to check mandatory Event fields before saving to DB
+    public Boolean eventIsOK(Event event) {
+        if(event.getStart() == null) return false;
+        if(event.getEnd() == null) return false;
+        if(event.getText() == null) return false;
+        if(event.getServDuration() == null) return false;
+
+        return true;
+    }
+
+    //Check if an event exist in the DB
+    public ResponseEntity<Boolean> eventExists(Long id) {
+        if(!eventRepo.existsById(id)) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }else {
+            return ResponseEntity.status(HttpStatus.OK).body(true);
+        }
     }
 
     public static class EventCreateParams {
